[Items] tile 클래스 분리
