+++ lib/Items.dart
@@ -9,6 +9,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_application/items_provider.dart';
 import 'package:flutter_application/refresh_fooddata.dart';
+import 'package:flutter_application/tile.dart';
 import 'package:provider/provider.dart';
 import 'firebase_options.dart';
 import 'package:flutter_application/main.dart';
@@ -265,61 +266,4 @@ Future<String> getFoodNameByFid(int fid) async {
 String getUserUid() {
   User? user = FirebaseAuth.instance.currentUser;
   return user?.uid ?? '';
-}
-
-class NewTile extends StatelessWidget {
-  final String? docId;
-  final String remainingDays;
-  final String foodName;
-  final int quantity;
-  final Function()? onDelete;
-  final Function()? onDecrease;
-
-  const NewTile({
-    required this.docId,
-    super.key,
-    required this.remainingDays,
-    required this.foodName,
-    required this.quantity,
-    this.onDelete,
-    this.onDecrease,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      elevation: 4,
-      child: ListTile(
-        title: Text(foodName),
-        leading: Text(
-          remainingDays,
-          style: const TextStyle(fontSize: 18),
-        ),
-        subtitle: Text('수량: $quantity'),
-        trailing: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            IconButton(
-              icon: const Icon(Icons.remove),
-              onPressed: quantity > 1 ? onDecrease : null,
-            ),
-            IconButton(
-              icon: const Icon(Icons.edit),
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) => EditFoodScreen(docId: docId!)),
-                );
-              },
-            ),
-            IconButton(
-              icon: const Icon(Icons.delete),
-              onPressed: onDelete,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
 }
